+++ test/next/jdbc/result_set_test.clj
@@ -95,13 +95,24 @@
       (is (nil? (:fruit/appearance row)))
       (is (= 3 (:fruit/id row)))
       (is (= "Peach" (:fruit/name row)))))
-  (testing "lower-case row builder"
+  (testing "unqualified lower-case row builder"
     (let [row (p/-execute-one (ds)
                               ["select * from fruit where id = ?" 4]
                               {:builder-fn rs/as-unqualified-lower-maps})]
       (is (map? row))
       (is (= 4 (:id row)))
-      (is (= "Orange" (:name row))))))
+      (is (= "Orange" (:name row)))))
+  (testing "custom row builder"
+    (let [row (p/-execute-one (ds)
+                              ["select * from fruit where id = ?" 3]
+                              {:builder-fn rs/as-modified-maps
+                               :label-fn str/lower-case
+                               :qualifier-fn identity})]
+      (is (map? row))
+      (is (contains? row :FRUIT/appearance))
+      (is (nil? (:FRUIT/appearance row)))
+      (is (= 3 (:FRUIT/id row)))
+      (is (= "Peach" (:FRUIT/name row))))))
 
 (deftest test-mapify
   (testing "no row builder is used"